requirements.txt
Pyexecjs
datetime
requests
configparser
